confluent-kafka-mongo-atlas/requirements.txt
avroconsumer
confluent-kafka
fastavro
requests
six
pandas
dnspython
pymongo
python-dotenv